+++ ACBrFramework.Net/ACBrFramework.Net/Sintegra/Registro/Registro 61R/SintegraRegistro61R.cs
@@ -2,16 +2,16 @@ namespace ACBrFramework.Sintegra
 {
 	public sealed class SintegraRegistro61R
 	{
-		public double Aliquota { get; set; }
+		public decimal Aliquota { get; set; }
 
-		public double Valor { get; set; }
+		public decimal Valor { get; set; }
 
-		public double Qtd { get; set; }
+		public decimal Qtd { get; set; }
 
 		public string MesAno { get; set; }
 
 		public string Codigo { get; set; }
 
-		public double BaseDeCalculo { get; set; }
+		public decimal BaseDeCalculo { get; set; }
 	}
 }
